fix: add form to component with page in playground
